Renamespace ServiceClient
